% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dna}
\alias{dna}
\title{Solution for Problem "Counting DNA Nucleotides"}
\usage{
dna(s)
}
\arguments{
\item{s}{A string holding the sequence.}
}
\value{
A vector with the number of A, C, G and T in that order.
}
\description{
Counts DNA nucleotides in sequence \code{s} and returns the number of A, C, G and T in that order. Letters other than A, C, G and T will be ignored.
}
