% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fib}
\alias{fib}
\title{Solution for Problem "Rabbits and Recurrence Relations"}
\usage{
fib(n, k)
}
\arguments{
\item{n}{The number of months as an integer.}

\item{k}{The number of rabbit pairs as an integer.}
}
\value{
The number of rabbit pairs after n months.
}
\description{
Calculates the number of rabbit pairs after n months.
}
