% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc}
\alias{gc}
\title{Solution for Problem "Computing GC Content"}
\usage{
gc(file_path)
}
\arguments{
\item{file_path}{The path of the FASTA file as a string.}
}
\value{
The highest GC percentage along with the name in the form of a named vector with one element.
}
\description{
Calculates the percentage of G and C nucleotides in the sequences of the given FASTA file and returns the highest GC percentage along with the name of the sequence as a named vector.
}
