% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hamm}
\alias{hamm}
\title{Solution for Problem "Counting Point Mutations"}
\usage{
hamm(s, t)
}
\arguments{
\item{s}{A DNA sequence as a string.}

\item{t}{A DNA sequence as a string.}
}
\value{
The number of point mutations.
}
\description{
Calculates the Hamming distance dH(s,t) of DNA strings s and t (i.e., the number of point mutations).
}
