% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_fasta}
\alias{parse_fasta}
\title{A simple FASTA-parser, implemented in C++}
\usage{
parse_fasta(file_path)
}
\arguments{
\item{file_path}{The path of a FASTA file.}
}
\value{
A named list with the contents of the FASTA file.
}
\description{
Parses a FASTA file and returns the content as a named list.
}
