% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{revc}
\alias{revc}
\title{Solution for Problem "Complementing a Strand of DNA"}
\usage{
revc(s)
}
\arguments{
\item{s}{A string holding the sequence.}
}
\value{
The complement, sc, of string s.
}
\description{
Reverses the sequence \code{s} and replaces each nucleotide with the complement (i.e., A<->T and G<->C). Non-nucleotide letters will be ignored.
}
