% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rna}
\alias{rna}
\title{Solution for Problem "Transcribing DNA into RNA"}
\usage{
rna(t)
}
\arguments{
\item{t}{A string holding the sequence.}
}
\value{
A string where each T has been replaced with an U.
}
\description{
Replaces each T with an U.
}
